Changes and build for Windows

+++ src/main/python/gui/components/AboutDialog.py
@@ -103,5 +103,5 @@ class AboutDialog(QDialog):
     def _center(self):
         """Centers the window on the screen"""
         screen = QDesktopWidget().screenGeometry()
-        self.move((screen.width() - self.geometry().width()) / 2,
-                  (screen.height() - self.geometry().height()) / 2)
+        self.move(int((screen.width() - self.geometry().width()) / 2),
+                  int((screen.height() - self.geometry().height()) / 2))
